Refactor: extract shared extraction/file-finding utils, fix encapsulation and ordering
- Move tar→PowerShell fallback extraction and recursive file search into
  downloadUtils.js to eliminate duplication across whisperCudaManager and
  llamaVulkanManager
- Convert sync readdirSync file search to async fsPromises.readdir
- Add isDownloading() getter to WhisperCudaManager instead of exposing
  _downloading directly
- Fix Vulkan download handler to persist env file before restarting server
