+++ src/helpers/downloadUtils.js
@@ -3,6 +3,7 @@ const { promises: fsPromises } = require("fs");
 const path = require("path");
 const https = require("https");
 const http = require("http");
+const { execFile } = require("child_process");
 const { pipeline } = require("stream");
 const debugLogger = require("./debugLogger");
 
@@ -402,10 +403,87 @@ async function checkDiskSpace(directory, requiredBytes) {
   }
 }
 
+function extractZipWindows(zipPath, destDir) {
+  return new Promise((resolve, reject) => {
+    execFile("tar", ["-xf", zipPath, "-C", destDir], (error) => {
+      if (error) {
+        debugLogger.info("tar extraction failed, trying PowerShell", { error: error.message });
+        execFile(
+          "powershell",
+          [
+            "-NoProfile",
+            "-Command",
+            `Expand-Archive -Force -Path '${zipPath}' -DestinationPath '${destDir}'`,
+          ],
+          (psError) => {
+            if (psError) reject(new Error(`Zip extraction failed: ${psError.message}`));
+            else resolve();
+          }
+        );
+      } else {
+        resolve();
+      }
+    });
+  });
+}
+
+function extractArchive(archivePath, destDir) {
+  if (archivePath.endsWith(".tar.gz") || archivePath.endsWith(".tgz")) {
+    return new Promise((resolve, reject) => {
+      execFile("tar", ["-xzf", archivePath, "-C", destDir], (err) => {
+        err ? reject(new Error(`Extraction failed: ${err.message}`)) : resolve();
+      });
+    });
+  }
+
+  if (process.platform === "win32") {
+    return extractZipWindows(archivePath, destDir);
+  }
+
+  return new Promise((resolve, reject) => {
+    execFile("unzip", ["-o", archivePath, "-d", destDir], (err) => {
+      err ? reject(new Error(`Extraction failed: ${err.message}`)) : resolve();
+    });
+  });
+}
+
+async function findFile(dir, name, maxDepth = 5, depth = 0) {
+  if (depth >= maxDepth) return null;
+  const entries = await fsPromises.readdir(dir, { withFileTypes: true });
+  for (const entry of entries) {
+    const full = path.join(dir, entry.name);
+    if (entry.isDirectory()) {
+      const found = await findFile(full, name, maxDepth, depth + 1);
+      if (found) return found;
+    } else if (entry.name === name) {
+      return full;
+    }
+  }
+  return null;
+}
+
+async function findFiles(dir, pattern, maxDepth = 5, depth = 0) {
+  if (depth >= maxDepth) return [];
+  const results = [];
+  const entries = await fsPromises.readdir(dir, { withFileTypes: true });
+  for (const entry of entries) {
+    const full = path.join(dir, entry.name);
+    if (entry.isDirectory()) {
+      results.push(...(await findFiles(full, pattern, maxDepth, depth + 1)));
+    } else if (pattern.test(entry.name)) {
+      results.push(full);
+    }
+  }
+  return results;
+}
+
 module.exports = {
   downloadFile,
   createDownloadSignal,
   validateFileSize,
   cleanupStaleDownloads,
   checkDiskSpace,
+  extractArchive,
+  findFile,
+  findFiles,
 };
